Require integer totalStarPlayerPoints in LineUpSchema
Star Player Points are always whole numbers. z.number() without
.int() would accept a fractional value from TP's JSON, which
UpsertPlayerSchema's own integer sppTotal would then reject at
upsert time -- failing far from where the bad data was read.
MatchLineUpSchema inherits the constraint since it derives from
LineUpSchema.

PR review feedback for #380.

+++ packages/parse-tp/src/roster-parser.service.ts
@@ -81,7 +81,7 @@ export const LineUpSchema = z.object({
   rosterId: z.number(),
   position: z.string(),
   isBigGuy: z.boolean().optional(),
-  totalStarPlayerPoints: z.number(),
+  totalStarPlayerPoints: z.number().int(),
 });
 
 const RosterSchema = z.object({
